reprint_forecast/requirements.txt
pandas==2.3.0
numpy==2.2.6
scikit-learn==1.6.1
xgboost==2.1.4
sqlalchemy==2.0.38
pyodbc==5.2.0
python-dotenv==1.1.0
pyyaml==6.0.2
joblib==1.4.2
